5 acceptance test

+++ app/controllers/questions_controller.rb
@@ -1,23 +1,22 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]  
 
   def index
     @questions = Question.all
-  end
-  
-  def show
-  end
-
-  def new
   end
 
   def edit
   end
 
+  def show
+    @answer = question.answers.new
+  end  
+
   def create
-    @question = Question.new(question_params)
-    
+    @question = current_user.questions.new(question_params)
+
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end  
@@ -32,8 +31,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: "Question successfully delete"
+    else
+      redirect_to question, notice: 'You are not the author question.'
+    end
   end
 
   private
